[feenkcom/gtoolkit#520] do not use `clipChildren:` as it significantly affects performance
Having `clipChildren: false` is very very expensive, because if an element does not clip children then to compute invalidation bounds we should iterate over its children and merge invalidation bounds together. Invalidation bounds is based on visual bounds which is based on stroked bounds which is very expensive to compute for a line

On other hand, not having the `clipChildren: false` line implies that dragging a Mondrian visualization will clip it at an unespected situation. Try the `GtMondrianBasicExamples new nestedWithEdges` example, drag the whole visualisation to the top right corner and then try to move one of the nodes to the bottom left corner, at some point that node will disappear if `clipChildren:` is `true`.